The tag table was the contract, written three times
`type-p` is where a `^int` or `^string` annotation stops being a number and
becomes a question. The numbers are the contract between the compiler and
every runtime, which is why all three copies wrote them out rather than
deriving them — and why three copies was the wrong number.

Its own comments record two drifts already repaired IN PLACE: a map entry had
to become a vector here as well as in `vector?`, and a row ref a map here as
well as in `map?`. The CLR's copy states the moral — "a switch that restates
predicates defined elsewhere will keep drifting from them; the fix is to stop
restating" — and then restates six of them, because at the time there was
nowhere else for the statement to live. There is now.

Two things the three copies did not agree on, both found by writing the port
rather than by a test:

* `fn?`. Rust answered from `is_fn`, which is closure, nativefn OR MULTIFN;
  the ports' arm 12 was closure or nativefn. LATENT, not live: nothing
  constructs a multifn — `defmulti` is a macro over a plain closure, and
  `TY_MULTIFN` is declared and unbuilt, like `TY_CHUNKSEQ`. Checked rather
  than assumed, by running the same program on both: `(fn? a-defmulti)` is
  true natively and true on the JVM, because the value is a closure either
  way. Converged toward Rust, which is what `kin/tablekind.kin` already said
  callable means.
* the `fn?` builtin's route. Rust called `is_fn`; the ports went through the
  table. One route now.

`is-sequential` came with it, being the same mistake one level down: Rust
asked `category`, both ports carried the tag list, and the JVM's copy once
read `is-vec or is-seq` — that list minus map entries — so
`(sequential? (first (seq m)))` was false there while the table two methods
above said true. It asks `category` now, in one place.

`test/types.clj` scraped the table out of `vm.rs` to check it against the
compiler's. It reads `kin/typep.kin` now, which is a STRONGER check and not
merely a moved one: the compiler's table against THE runtime table rather
than against whichever copy the scraper could find, with `bin/check-kin`
holding the three emissions to it. Mutation-verified in both directions — a
renamed function and a wrong arm each turn it red.

One caller was missed by `cargo check` and caught by `bin/build-dist`:
`aot.rs` is behind a feature the default check does not compile. Its cast is
gone rather than widened — the AOT ABI already passes a u32.

bin/test, conform-hosts (249 checks), check-kin (76 sources), check-names,
check-builtins, check-dist, check-decisions, check-flattens: all green.

+++ runtime/src/aot.rs
@@ -222,7 +222,7 @@ pub extern "C" fn aot_int_binop(
 pub extern "C" fn aot_type_p(rt: *mut Rt, code: u32, top: u32) -> u64 {
     let rt = unsafe { &mut *rt };
     let v = rt.roots.stack[top as usize - 1];
-    Value::boolean(rt.type_p(code as u8, v)).0
+    Value::boolean(rt.type_p(code, v)).0
 }
 
 /// `RETURN`. Pops the frame and pushes the result where the caller expects it,
